fix some stuff

+++ src/main/java/at/technikumwien/repositories/DocumentRepository.java
@@ -3,5 +3,6 @@ package at.technikumwien.repositories;
 import at.technikumwien.entities.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DocumentRepository extends JpaRepository<Document, Integer> {
+public interface DocumentRepository extends JpaRepository<Document, Long> {
+
 }
